+++ js/shared/facebook-admin-reporter.js
@@ -740,3 +740,4 @@
     };
 }(typeof globalThis !== 'undefined' ? globalThis : window));
 
+